src/bookshelf/syncer_v2/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
Pillow>=10.0.0
requests==2.32.3
urllib3==2.3.0
